# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nebula Browser API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ProxyFetchParams = {
  /**
   * The URL to proxy
   */
  url: string;
};

export type ProxyResourceParams = {
  /**
   * The resource URL to proxy
   */
  url: string;
};