# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nebula Browser API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: proxy
    description: Web proxy operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /proxy:
    get:
      operationId: proxyFetch
      tags: [proxy]
      summary: Proxy a URL
      description: Fetches and proxies web content through the server
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: The URL to proxy
      responses:
        "200":
          description: Proxied HTML content
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proxy/resource:
    get:
      operationId: proxyResource
      tags: [proxy]
      summary: Proxy a resource
      description: Fetches and proxies a resource (CSS, JS, images) through the server
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: The resource URL to proxy
      responses:
        "200":
          description: Proxied resource
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message